pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bookmark_enrichment_tool"
version = "0.1.0"
description = "A Python application that processes CSV bookmark exports and enriches them with AI-generated content summaries and tags"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Bookmark Enrichment Tool", email = "dev@example.com"},
]
keywords = ["bookmarks", "ai", "content-analysis", "web-scraping"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "pandas>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "newspaper3k>=0.2.8",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0.0",
    "lxml[html_clean]>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
bookmark-enricher = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
include = [
    "main.py",
    "config.py", 
    "config.yaml",
    "src/**/*.py"
]

[project.urls]
Homepage = "https://github.com/example/bookmark-enrichment-tool"
Repository = "https://github.com/example/bookmark-enrichment-tool"
Issues = "https://github.com/example/bookmark-enrichment-tool/issues"

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]